more 32 bit cost safety

+++ include/osr/routing/astar.h
@@ -172,7 +172,7 @@ struct astar {
               neighbor.print(std::cout, w);
             }
 
-            auto const total = curr_cost + cost;
+            auto const total = static_cast<std::uint64_t>(curr_cost) + cost;
             auto const heur =
                 total + heuristic(params, w, sharing, neighbor.get_node());
             if (total >= max) {
